requirements.txt
streamlit==1.28.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
pillow==10.0.0
jupyter==1.0.0
ipywidgets==8.1.0
